IRBuilder: ensure names, entity header, consts.
Ensure that "structs" have names, so that structs can be referred
symbolically. In the future, functions and basic blocks also need to be
ensured to have names. Such names are generated from their IDs if not
provided by the client.

The IR builder can create named or unnamed MuEntityHeader based on
whether it has client-provided name or not.

Can build constants now (currently only Constant::Int).